requirements.txt
langchain==0.1.12
langchain-core>=0.1.31
langchain-community==0.0.28
python-dotenv==1.0.1
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pypdf==3.17.1
docx2txt==0.8
chromadb==0.4.22
tiktoken==0.6.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pydantic==2.6.1
numpy==1.26.4
pandas==2.2.0
cryptography>=3.1
reportlab==4.1.0 